% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCodingmarkers.R
\name{addDecision}
\alias{addDecision}
\title{addDecision}
\usage{
addDecision(
  dt,
  metadata_path = "../data/SAMPLE_INFO.txt",
  trim = FALSE,
  sample_column = 4,
  target_column = 23
)
}
\arguments{
\item{dt}{A data frame containing decision table. The last column of decision should be missing at this point. For experts, this is an information system.}

\item{metadata_path}{A string pointing to the metadta file in your system. for e.g. '../data/SAMPLE_INFO.txt'}

\item{trim}{A boolean value to be set to TRUE by people using St Judes dataset. By default is set to FALSE}

\item{sample_column}{A integer value containing the column of the sample names in the metadata}

\item{target_column}{A integer value containing the column of the decision value to be appended to the decision table.}
}
\value{
A dataframe which is a true decision table with decision in the last column and could be used in ML applications.
}
\description{
Remove non-coding, pseudo, rRNAs, scRNA, snoRNA, snRNA, unknown and other genes based on org.Hs.eg.db
}
\examples{
# dt<- addDecision(dt, metadata_path="../data/SAMPLE_INFO.txt",
# trim=FALSE, sample_column=4 ,target_column=23)
}
\author{
Girish Pulinkala
}
