% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSETTA-Ex.R
\name{boost_check_case}
\alias{boost_check_case}
\title{boost_check_case}
\usage{
boost_check_case(mcfs_result, col_size)
}
\arguments{
\item{mcfs_result}{A dataframe consiting of RI values from mcfs run.}

\item{col_size}{An integer value containng the nummber  in df}
}
\value{
A integer specifying which case to run for boosting
}
\description{
checks if the mcfs result is worth boosting
}
\author{
Girish Pulinkala
}
