% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSETTA-Ex.R
\name{boost_ros}
\alias{boost_ros}
\title{quality_check}
\usage{
boost_ros(
  ros_data,
  step = 2,
  discrete = FALSE,
  reducer = "Johnson",
  JohnsonParam = list(Modulo = TRUE, BRT = FALSE, BRTprec = 0.9, Precompute = FALSE,
    Approximate = TRUE, Fraction = 0.95),
  GeneticParam = list(Modulo = TRUE, BRT = FALSE, BRTprec = 0.9, Precompute = FALSE,
    Approximate = TRUE, Fraction = 0.95, Algorithm = "Simple")
)
}
\arguments{
\item{ros}{A output from the rosetta function}

\item{quality}{A vector containing quality of all models from boosting}
}
\value{
A boolean vaue inicating whether the current run was the best run.
}
\description{
does a quality check to find the accuracte or best model.
}
\author{
Girish Pulinkala
}
