% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCFS.R
\name{check_log_file}
\alias{check_log_file}
\title{check_log_file}
\usage{
check_log_file(mcfs_out_path)
}
\arguments{
\item{mcfs_out_path}{A file path to save mcfs files}
}
\description{
Checks if a log file exists in the folder
}
\author{
Girish Pulinkala
}
