% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCFS.R
\name{create_indexed_file}
\alias{create_indexed_file}
\title{create_indexed_file}
\usage{
create_indexed_file(filename, extension = ".txt", file_path)
}
\arguments{
\item{filename}{A input from check_log_file}

\item{extension}{A input from check_log_file}

\item{file_path}{A input from check_log_file}
}
\description{
creates indexed log file if a log file exists
}
\author{
Girish Pulinkala
}
