% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeGenenames-Ex.R
\name{data_prep}
\alias{data_prep}
\title{create_indexed_file}
\usage{
data_prep(data, transpose = TRUE)
}
\arguments{
\item{filename}{A input from check_log_file}

\item{extension}{A input from check_log_file}

\item{file_path}{A input from check_log_file}
}
\description{
creates indexed log file if a log file exists
}
\author{
Girish Pulinkala
}
