% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysis.R
\name{DEAnalysis}
\alias{DEAnalysis}
\title{DEAnalysis from a De cision table}
\usage{
DEAnalysis(dt, keep_val = 20, compare = "iALL")
}
\arguments{
\item{dt}{A Decision table}

\item{keep_val}{A integer  which specifies the rowMeans value}

\item{compare}{A string from the binary outcome that you want to compare with}
}
\description{
generates DE analysis for GE data
}
\author{
Girish Pulinkala
}
