% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCodingmarkers.R
\name{findCodingmarkers}
\alias{findCodingmarkers}
\title{findCodingmarkers}
\usage{
findCodingmarkers(dt, keytype = "SYMBOL")
}
\arguments{
\item{dt}{A data frame containing decision table. The last column of decision should be missing at this point. For experts, this is an information system.}

\item{keytype}{the keytype that matches the keys used. For the select methods, this is used to indicate the kind of ID being used with the keys argument. For the keys method this is used to indicate which kind of keys are desired from keys}
}
\value{
A dataframe with only protein coding regions, if transposed can be used for differnetial expressions too
}
\description{
Find protein coding and trim the dataframe from all other or noncoding regions
}
\examples{
# path <- c("../data/data_all_old/")
# dt<- readFeaturecounts(path,makedt=TRUE)
# dt<-findCodingmarkers(dt,keytype="SYMBOL")
}
\author{
Girish Pulinkala
}
