% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSETTA-Ex.R
\name{mcfs_to_rosetta}
\alias{mcfs_to_rosetta}
\title{mcfs_to_rosetta}
\usage{
mcfs_to_rosetta(mcfs_df, out_name = NULL)
}
\arguments{
\item{ros}{A output from the rosetta function}

\item{quality}{A vector containing quality of all models from boosting}
}
\value{
A boolean vaue inicating whether the current run was the best run.
}
\description{
Does feture boosting and builts rule-based models from mcfs input
}
\author{
Girish Pulinkala
}
