% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSETTA-Ex.R
\name{quality_check}
\alias{quality_check}
\title{quality_check}
\usage{
quality_check(ros, quality)
}
\arguments{
\item{ros}{A output from the rosetta function}

\item{quality}{A vector containing quality of all models from boosting}
}
\value{
A boolean vaue inicating whether the current run was the best run.
}
\description{
does a quality check to find the accuracte or best model.
}
\author{
Girish Pulinkala
}
