% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFeaturecounts.R
\name{readFeaturecounts}
\alias{readFeaturecounts}
\title{readFeaturecounts}
\usage{
readFeaturecounts(folderpath, makedt = TRUE)
}
\arguments{
\item{folderpath}{A string value pointing to the folder containing read counts for each sample individually for .e.g. SJACT001_D.RNA-Seq.feature-counts.txt}

\item{makedt}{A boolean value that specifies whether to generate a decision table or a dataframe for differential expression.}
}
\value{
A dataframe for ML or Differential expression analysis
}
\description{
Make a Decision table from htseq read counts,The last column of decision should be missing at this point. For experts, this is an information system.
}
\examples{
# path <- c("../data/data_all_old/")
# df<- readFeaturecounts(path,makedt=TRUE)
}
\author{
Girish Pulinkala
}
