% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCodingmarkers.R
\name{removeRPgenes}
\alias{removeRPgenes}
\title{removeRPgenes}
\usage{
removeRPgenes(dt)
}
\arguments{
\item{dt}{A data frame containing decision table. The last column of decision should be missing at this point. For experts, this is an information system.}
}
\value{
A dataframe with no Ribosomal protein coding genes.
}
\description{
Remove ribosomal protein coding genes from the decision table based on a regular expression value.
}
\examples{
# path <- c("../data/data_all_old/")
# dt<- readFeaturecounts(path,makedt=TRUE)
# dt<-findCodingmarkers(dt,keytype="SYMBOL")
# dt <- removeRPgenes(dt)
}
\author{
Girish Pulinkala
}
