% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCFS.R
\name{runMCFS}
\alias{runMCFS}
\title{runMCFS}
\usage{
runMCFS(
  mcfs_df,
  out_name,
  permutations = 3,
  set_response_var = FALSE,
  response_var = NULL,
  plot_res = TRUE,
  projections = "auto",
  projectionSize = "auto",
  balance = 2
)
}
\arguments{
\item{mcfs_df}{A dataframe or a decision table with decision as the last column.}

\item{out_name}{A string filename for the final mcfs.rds file, for e.g. 'autcon'}

\item{permutations}{A integer value which specfies the number of permutations It is from the original mcfs function.}

\item{set_response_var}{A boolean value, which says if the user wamts to specifiy the decision column. by default it is set to FALSE. Use this if the your last column is not the decision column}

\item{response_var}{A string value to specify the decision column name. By default is set to FALSE and will automatically set the last column to be decision column}

\item{plot_res}{A boolean value that sepcifies whether to plot the graphs from the mcfs runs. By deafult it is set to TRUE.}

\item{projections}{An integer value specifying number of projections. Refer to rmcfs documentation.}

\item{projectionSize}{An integer value specifying number of projectionSize. Refer to rmcfs documentation.}

\item{balance}{An integer value to specify undersampling. Refer to rmcfs documentation.}
}
\description{
Runs Monte Carlo Feature selection and saves outputs and logs
}
\examples{
# example code
# dt <- read.delim('alizadeh.csv)
# runMCFS(dt, 'mcfs_run1',permutations=20,set_response_var=FALSE,response_var=NULL,plot_res=TRUE)
}
\author{
Girish Pulinkala
}
