% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSETTA-Ex.R
\name{select_cols_boost}
\alias{select_cols_boost}
\title{select_cols_boost}
\usage{
select_cols_boost(boost_case_num, col_size, mcfs_result)
}
\arguments{
\item{boost_case_num}{The output from the boost_check_case function}

\item{col_size}{An integer value containng the nummber  in df}

\item{mcfs_result}{A dataframe consiting of RI values from mcfs run.}
}
\value{
A vector of values containing columns that can be used to train your model.
}
\description{
Selects the number of columns to boost. Since r.roseeta has a limit it selects a maximu of 500 columns
}
\author{
Girish Pulinkala
}
