% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysis.R
\name{TMMnormalisation}
\alias{TMMnormalisation}
\title{TMMnormalization}
\usage{
TMMnormalisation(dt)
}
\arguments{
\item{dt}{A decision table}
}
\value{
A dataframe for ML
}
\description{
TMM normalisation for decision system
}
\examples{
# norm_dt<- TMMnormalisation(dt)
}
\author{
Girish Pulinkala
}
